don't double-free OPs on blead [#83439]
Starting with 5.17.2 all newly allocated OPs are implicitly registered
with the current CV being compiled (PL_compcv). If an error is thrown
during compilation, these OPs are freed with the CV. This results in a
double-free if we also track and free OPs ourselves.

Solution: Check for this case and don't call op_free() in this
situation.