don't predeclare subs that have a custom install_sub (RT #120804)
We don't know what the provided install_sub is going to do with the
coderef so it makes no sense to tell the perl compiler it's going to be
in the symbol table.

Also, Pkg->can("foo") returns true for 'sub foo;' stubs, which would end
up breaking Class::Method::Modifiers.